requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
Pillow>=11.0.0
python-pptx==0.6.23
python-multipart==0.0.9
beautifulsoup4==4.12.3
cairosvg==2.9.0
